fix warnings (std::size_t to int conversions), MAD : 1 -> 1.5,

+++ Surface_mesh_segmentation/include/CGAL/internal/Surface_mesh_segmentation/Expectation_maximization.h
@@ -220,7 +220,7 @@ private:
                 if(distance < min_distance)
                 {
                     min_distance = distance;
-                    closest_center = i;
+                    closest_center = static_cast<int>(i);
                 }
             }
             member_count[closest_center]++;
